requirements.txt
Django==2.2
entrypoints==0.3
flake8==3.7.8
mccabe==0.6.1
psycopg2==2.8.4
pycodestyle==2.5.0
pyflakes==2.1.1
pytz==2019.3
sqlparse==0.3.0
django-reversion==3.0.7